refactor(vault): decompose FolderManager into focused sub-components
- FolderManager.tsx: 1,252 → 40 lines (97% reduction)
- Created folder-manager/ subdirectory with 5 files:
  - FolderManagerTypes.ts (85L) - all type definitions
  - FolderManagerIcons.tsx (299L) - 14 SVG icons
  - FolderList.tsx (210L) - tree display + buildFolderTree helper
  - FolderDialogs.tsx (397L) - create/edit dialogs
  - FolderPanels.tsx (333L) - share/items panels + helpers
  - index.ts (19L) - barrel exports
- Namespace export preserved via aliased imports
- All 36 characterization tests pass
- Zero TypeScript errors, build successful
