+++ src/components/vault/folder-manager/index.ts
@@ -1,0 +1,19 @@
+/**
+ * Folder Manager - Barrel Exports
+ *
+ * Re-exports all folder management components and types.
+ */
+
+// Types
+export type {
+  FolderListProps,
+  FolderCreateDialogProps,
+  FolderEditDialogProps,
+  FolderSharePanelProps,
+  FolderItemsPanelProps,
+} from './FolderManagerTypes';
+
+// Components
+export { FolderList } from './FolderList';
+export { FolderCreateDialog, FolderEditDialog } from './FolderDialogs';
+export { FolderSharePanel, FolderItemsPanel } from './FolderPanels';
